+++ modules/DS18B20/DS18B20.cpp
@@ -1,3 +1,4 @@
+#include "../../settings.h"
 #ifdef MODULE_DS18B20
 
 #include "DS18B20.h"